+++ sso-glide/src/main/java/it/niedermann/nextcloud/sso/glide/StringLoader.kt
@@ -1,7 +1,6 @@
 package it.niedermann.nextcloud.sso.glide
 
 import android.content.Context
-import android.util.Log
 import com.bumptech.glide.load.Options
 import com.bumptech.glide.load.model.ModelLoader
 import com.bumptech.glide.load.model.ModelLoader.LoadData
@@ -15,28 +14,20 @@ import com.nextcloud.android.sso.model.SingleSignOnAccount
 import java.io.InputStream
 
 /**
- * A simple model loader for fetching media over http/https using OkHttp.
+ * Responsible for trying to fetch resources without an explicit [SingleSignOnAccount].
+ * It will use the currently set [SingleSignOnAccount] of [SingleAccountHelper] if possible and also handle relative URLs.
  */
 class StringLoader(private val context: Context) : ModelLoader<String, InputStream> {
-    private val TAG = StringLoader::class.java.simpleName
 
     override fun handles(url: String): Boolean {
-        Log.i(TAG, "[handles] ------------------------------")
-        Log.i(TAG, "[handles] ${url}")
         return try {
-            Log.i(TAG, "[handles] starts with / → ${url.startsWith("/")}")
-            Log.i(TAG, "[handles] starts with ${SingleAccountHelper.getCurrentSingleSignOnAccount(context).url} → ${url.startsWith(SingleAccountHelper.getCurrentSingleSignOnAccount(context).url)}")
-
             // We support this url if it starts with the current SingleSignOn accounts url
             url.startsWith(SingleAccountHelper.getCurrentSingleSignOnAccount(context).url) ||
                     // We also try to handle relative paths, assuming the requested resource is located at the current SingleSignOn account
                     url.startsWith("/")
-
         } catch (e: NextcloudFilesAppAccountNotFoundException) {
-            Log.i(TAG, "[handles] NextcloudFilesAppAccountNotFoundException → true")
             false
         } catch (e: NoCurrentAccountSelectedException) {
-            Log.i(TAG, "[handles] NoCurrentAccountSelectedException → true")
             false
         }
     }
@@ -49,11 +40,9 @@ class StringLoader(private val context: Context) : ModelLoader<String, InputStre
         })
     }
 
-    /**
-     * The default factory for [StringLoader]s.
-     */
     class Factory(context: Context) : ModelLoaderFactory<String, InputStream> {
         private val loader: StringLoader = StringLoader(context)
+
         override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<String, InputStream> {
             return loader
         }
@@ -61,6 +50,5 @@ class StringLoader(private val context: Context) : ModelLoader<String, InputStre
         override fun teardown() {
             // Do nothing, this instance doesn't own the client.
         }
-
     }
 }
